# Optimise isReserved