This function is called many thousands of times by ghcide and was a
hotspot in profiling.

This patch reduced time taken on one benchmark which called hover 1000
times from 259 seconds to 170 seconds.

See https://github.com/digital-asset/ghcide/issues/101