@fendor 2020-01-26
Maybe it makes sense to add this information as a comment? So that feature devs understand why this optimisation has been done?