@AndreasPK 2020-01-27
mpickering showed this patch off to me so I look at the code a bit.

`isReserved` is really the meat of what this patches optimizes.

Before this patch it compiled to this core:

```haskell
  = \ (c1_abyn :: Char) ->
      case GHC.List.elem
             @ Char
             ghc-prim-0.5.3:GHC.Classes.$fEqChar
             c1_abyn
             Network.URI.isAllowedInURI5
      of {
        False ->
          GHC.List.elem
            @ Char
            ghc-prim-0.5.3:GHC.Classes.$fEqChar
            c1_abyn
            Network.URI.isAllowedInURI3;
        True -> ghc-prim-0.5.3:GHC.Types.True
      }
```

essentially elem c foo || elem c bar

With this patch isReserved compiles to this:
```haskell
Network.URI.$wisReserved
  = \ (ww_sq86 :: ghc-prim-0.5.3:GHC.Prim.Char#) ->
      case ww_sq86 of {
        __DEFAULT -> ghc-prim-0.5.3:GHC.Types.False;
        '!'# -> ghc-prim-0.5.3:GHC.Types.True;
        '#'# -> ghc-prim-0.5.3:GHC.Types.True;
        '$'# -> ghc-prim-0.5.3:GHC.Types.True;
        '&'# -> ghc-prim-0.5.3:GHC.Types.True;
        '\''# -> ghc-prim-0.5.3:GHC.Types.True;
        '('# -> ghc-prim-0.5.3:GHC.Types.True;
        ')'# -> ghc-prim-0.5.3:GHC.Types.True;
        '*'# -> ghc-prim-0.5.3:GHC.Types.True;
        '+'# -> ghc-prim-0.5.3:GHC.Types.True;
        ','# -> ghc-prim-0.5.3:GHC.Types.True;
        '/'# -> ghc-prim-0.5.3:GHC.Types.True;
        ':'# -> ghc-prim-0.5.3:GHC.Types.True;
        ';'# -> ghc-prim-0.5.3:GHC.Types.True;
        '='# -> ghc-prim-0.5.3:GHC.Types.True;
        '?'# -> ghc-prim-0.5.3:GHC.Types.True;
        '@'# -> ghc-prim-0.5.3:GHC.Types.True;
        '['# -> ghc-prim-0.5.3:GHC.Types.True;
        ']'# -> ghc-prim-0.5.3:GHC.Types.True
      }
```

The former has overhead of two function calls, and the actual checking loops.

The later has no overhead + results in a binary search for the pattern match.

As a result the later is about 10 times as fast in a simple benchmark like this:

```haskell
      env (return "http://ex.it/foo/bar/baz/bap") $ \u1 ->
        bench "u1" $ nf (map (isReserved)) u1,
```

Currently processing one character like this takes somewhere between 100 and 150ns.

After this patch it takes 10ns and 15ns.  
That's only about 60 cycles and includes traversing one element of the list so not bad.