@osa1 2020-01-27
Amazing.

> Maybe it makes sense to add this information as a comment?

Agreed, someone in a few months will look at this code and say "why not use `elem`?" and revert this change.