@osa1 2020-01-27
This may be a bug in `elem/build` rule in GHC, see https://gitlab.haskell.org/ghc/ghc/issues/17752#note_250120