@ezrakilty 2020-01-27
It would be great if the compiler did the right thing for the `elem` version. In the meantime, though, I don't mind this being written out as a case match. I'll add a comment explaining why we've done so and linking back to here.